fix: separating concerns between model, controller, util-functions and creating authService which contains signUpUser
